+++ src/components/StopWatchButton.tsx
